project2/requirements.txt
httpx==0.27.2
python-dotenv==1.0.0

